typst/generate_matrix: init, allow column-wise input
Found a much simpler version on Github. Who needs replacement nodes? We
also add a feature I've been wanting for a while. Before, the snippet
placeholders would always be by-row:

mat(
  $1, $2;
  $3, $4;
)

But now, I have a mode to make them be by-column:

mat(
  $1, $3;
  $2, $4;
)

When I'm encoding a set of basis vectors as a matrix, this is far easier
than going per-row. And thanks to some smart querying logic, it doesn't
even muddy the code that much!